refactor: 패키지 이름 변경 및 폴더 이동
- user -> member 패키지 이름 변경
- Dto 패키지 분리
