+++ src/main/java/com/woowa/cafe/dto/member/UpdateMemberDto.java
@@ -1,4 +1,4 @@
-package com.woowa.cafe.dto;
+package com.woowa.cafe.dto.member;
 
 import com.woowa.cafe.domain.Member;
 
@@ -7,10 +7,10 @@ import java.util.Map;
 public record UpdateMemberDto(String lastPassword, String newPassword, String name, String email) {
     public static UpdateMemberDto from(final Map<String, String> params) {
         return new UpdateMemberDto(
-            params.get("lastPassword"),
-            params.get("newPassword"),
-            params.get("name"),
-            params.get("email")
+                params.get("lastPassword"),
+                params.get("newPassword"),
+                params.get("name"),
+                params.get("email")
         );
     }
