+++ src/main/java/com/woowa/cafe/service/ArticleService.java
@@ -2,9 +2,9 @@ package com.woowa.cafe.service;
 
 import com.woowa.cafe.domain.Article;
 import com.woowa.cafe.domain.Member;
-import com.woowa.cafe.dto.SaveArticleDto;
+import com.woowa.cafe.dto.article.SaveArticleDto;
 import com.woowa.cafe.repository.qna.ArticleRepository;
-import com.woowa.cafe.repository.user.MemberRepository;
+import com.woowa.cafe.repository.member.MemberRepository;
 
 import java.util.List;
