+++ src/main/java/com/woowa/cafe/service/MemberService.java
@@ -1,9 +1,9 @@
 package com.woowa.cafe.service;
 
 import com.woowa.cafe.domain.Member;
-import com.woowa.cafe.dto.SaveMemberDto;
-import com.woowa.cafe.dto.UpdateMemberDto;
-import com.woowa.cafe.repository.user.MemberRepository;
+import com.woowa.cafe.dto.member.SaveMemberDto;
+import com.woowa.cafe.dto.member.UpdateMemberDto;
+import com.woowa.cafe.repository.member.MemberRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
